code (Yelp API)/requirements.txt
requests==2.27.1
SQLAlchemy==1.4.34
python-dotenv==0.20.0
psycopg2-binary==2.9.3